requirements.txt
streamlit~=1.34.0
tensorflow~=2.16.1
pillow~=10.3.0
numpy~=1.26.4
pandas~=2.2.2
matplotlib~=3.9.1
